backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0
alembic==1.13.3
python-dotenv==1.0.1
openai==1.54.0
PyPDF2==3.0.1
